在移动文件的时候添加延迟，减少失误

+++ office_automation.py
@@ -7,8 +7,9 @@ import color_div.color_divider
 import re
 import os
 import random
+import time
 
-def done(func):
+def done(func):             # 装饰器
     def wrapper(*args, **kwargs):
         print('processing...')
         func(*args, **kwargs)
@@ -16,7 +17,9 @@ def done(func):
     return wrapper
 
 @done
-def main(*num, path, ifcolordiv=True):       # num是不需要移动到主图文件夹的图片序号,类似于(1),(2),(3)
+def main(path, ifcolordiv=True):
+    """path：工作路径
+    ifcolordiv：是否需要颜色分类"""
     list_path = renamer.main1.get_file_path(path)
     img1200_path = []
     img_path = []
@@ -46,8 +49,10 @@ def main(*num, path, ifcolordiv=True):       # num是不需要移动到主图文
 
     # 将1200图片拷贝到主图文件夹
     try:
-        for i in random.choices(img1200_path, k=5):
+        list_random = random.sample(img1200_path, k=5)
+        for i in list_random:
             shutil.copy(i, os.path.join(path, '主图'))
+            time.sleep(0.1)
     except ValueError:
         for i in img1200_path:
             shutil.copy(i, os.path.join(path, '主图'))
@@ -70,9 +75,11 @@ def main(*num, path, ifcolordiv=True):       # num是不需要移动到主图文
     check_path = renamer.main1.get_file_path(os.path.join(path, '主图'))
     list_ = ['1', '2', '3', '4', '5']
     for i in check_path:
-        num__ = re.search(r'\((\d+)\)', i).group(1)
-        if num__ not in list_:
-            os.remove(i)
+        match = re.search(r'\((\d+)\)', i)
+        if match:
+            num__ = match.group(1)
+            if num__ not in list_:
+                os.remove(i)
     # endregion
 
     if ifcolordiv:          # 颜色分类
